SNAP-345: Get Dora (Elasticsearch) average response time for single user under 250ms
