+++ dora-api/src/main/java/gov/ca/cwds/inject/EsRestClientManager.java
@@ -1,8 +1,9 @@
-package gov.ca.cwds.rest;
+package gov.ca.cwds.inject;
 
 import static gov.ca.cwds.dora.DoraUtils.createElasticsearchClient;
 
 import com.google.inject.Inject;
+import gov.ca.cwds.rest.ElasticsearchConfiguration;
 import gov.ca.cwds.rest.api.DoraException;
 import io.dropwizard.lifecycle.Managed;
 import java.io.IOException;
@@ -11,7 +12,7 @@ import org.elasticsearch.client.RestClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public final class EsRestClientManager implements Managed {
+public class EsRestClientManager implements Managed {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(EsRestClientManager.class);
 
@@ -31,11 +32,12 @@ public final class EsRestClientManager implements Managed {
   @Override
   public void stop() {
     try {
+      LOGGER.debug("\n********* EsRestClientManager stop is invoked *********");
       if (esRestClient != null) {
         esRestClient.close();
         esRestClient = null;
+        LOGGER.info("\n********* Elasticsearch client is closed *********");
       }
-      LOGGER.info("\n********* Elasticsearch client is closed *********");
     } catch (IOException e) {
       LOGGER.error("error closing Elasticsearch client", e);
       throw new DoraException("error closing Elasticsearch client", e);
